+++ xenon-common/src/main/java/com/vmware/xenon/common/Utils.java
@@ -717,7 +717,7 @@ public final class Utils {
         if (useBinary && source.getAction() != Action.POST) {
             try {
                 byte[] encodedBody = Utils.encodeBody(source, source.getLinkedState(),
-                        Operation.MEDIA_TYPE_APPLICATION_KRYO_OCTET_STREAM);
+                        Operation.MEDIA_TYPE_APPLICATION_KRYO_OCTET_STREAM, true);
                 source.linkSerializedState(encodedBody);
             } catch (Throwable e2) {
                 Utils.logWarning("Failure binary serializing, will fallback to JSON: %s",
@@ -734,11 +734,11 @@ public final class Utils {
         }
     }
 
-    public static byte[] encodeBody(Operation op) throws Throwable {
-        return encodeBody(op, op.getBodyRaw(), op.getContentType());
+    public static byte[] encodeBody(Operation op, boolean isRequest) throws Throwable {
+        return encodeBody(op, op.getBodyRaw(), op.getContentType(), isRequest);
     }
 
-    public static byte[] encodeBody(Operation op, Object body, String contentType)
+    public static byte[] encodeBody(Operation op, Object body, String contentType, boolean isRequest)
             throws Throwable {
         if (body == null) {
             op.setContentLength(0);
@@ -777,7 +777,31 @@ public final class Utils {
             op.setContentLength(data.length);
         }
 
+        // For requests, if encoding is specified as gzip, then compress body
+        // For responses, if request accepts gzip body, then compress body and add response header
+        String encodingHeader = isRequest ?
+                Operation.CONTENT_ENCODING_HEADER : Operation.ACCEPT_ENCODING_HEADER;
+        String encoding = op.getRequestHeaderAsIs(encodingHeader);
+        if (Operation.CONTENT_ENCODING_GZIP.equals(encoding)) {
+            data = compressGZip(data);
+            op.setContentLength(data.length);
+            if (!isRequest) {
+                op.addResponseHeader(Operation.CONTENT_ENCODING_HEADER, Operation.CONTENT_ENCODING_GZIP);
+            }
+        }
+
         return data;
+    }
+
+   /**
+     * Compresses byte[] to gzip byte[]
+     */
+    private static byte[] compressGZip(byte[] input) throws Exception {
+        ByteArrayOutputStream out = new ByteArrayOutputStream();
+        try (GZIPOutputStream zos = new GZIPOutputStream(out)) {
+            zos.write(input, 0, input.length);
+        }
+        return out.toByteArray();
     }
 
     /**